requirements.txt
accelerate>=0.33.0
deepspeed>=0.15.0
ms-swift>=3.4.0
transformers>=4.53.0
